pyproject.toml
[project]
name = "scientific-rss-reader"
version = "0.1.0"
description = "Async RSS -> Telegram channel notifier demo (uv)"
requires-python = ">=3.10"
dependencies = [
  "aiohttp>=3.9.5",
  "aiosqlite>=0.20.0",
  "feedparser>=6.0.11",
  "python-dotenv>=1.0.1",
]

[tool.uv]
